refactor(storage): extract me-index crate (graph + scope projections), Wave 2 S2 (#930)
Carve the in-memory projections `MemoryGraph` (petgraph) + `ScopeTree` out of the
`memory-engine` facade into a new backend-free L2 leaf crate `me-index`, depending
only on `me-types`. First of three S2 sub-PRs (me-index -> me-backend-sqlite ->
me-backend-postgres). The facade re-exports `crate::{graph,scope}` as `pub(crate)`
(their pre-carve visibility), so the consumer-facing public API is unchanged.

Breaking the graph/scope <-> store cycle — invisible in the monolith (module cycles
are legal within one crate), a hard error once split into sibling crates:

- `validate_segment` + `MAX_SEGMENT_LEN` -> `me-types` (the shared scope-segment SSOT;
  breaks store->index). Both the SQLite write path (`ScopeStore::ensure_path`) and the
  index read path (`ScopeTree::resolve_path`) now depend on it downward.
- The two conn-based loaders (`MemoryGraph::load_from_db`, `ScopeTree::load` — one
  production caller each: the engine open-path rebuild fallback) -> facade free
  functions in `engine/graph_load.rs` (breaks index->store). `me-index` keeps only the
  pure cores (`from_active_edges`, `from_snapshot`) plus a new `ScopeTree::from_nodes`.
  This sheds `rusqlite` from `me-index` entirely — `cargo tree -p me-index` is
  `me-types` + `petgraph` only.
- Their DB-touching tests relocate with them (4 `load_from_db_*` tests + `setup_db` + 2
  scope round-trip proptests -> `graph_load.rs`); scope's ~25 unit tests get a DB-free
  `setup_tree` built via `from_nodes`. Net test-fn delta: 0 (1967 pass).

Forced cross-crate visibility widenings (`me-index` is `publish=false` = workspace
internal; the facade re-exports `pub(crate)`, so these are NOT consumer-facing):
`to_snapshot`/`from_snapshot` `pub(crate)`->`pub`; `#[must_use]` on the now
genuinely-public `ScopeTree` query methods. An `archive` marker feature on `me-index`
(gates `MemoryGraph::remove_node`) is forwarded from the facade's `archive` feature so
they compile in lockstep.

Full CI-exact §I gate green (fmt, clippy `-D warnings`, build x2, test `--all-features`
1967/0, doc x2, doc me-index, MSRV x2). Reviewed via rust-specialist carve + an
independent re-gate.
