+++ Cargo.toml
@@ -4,6 +4,7 @@ members = [
     "memory-engine/lib/me-types",
     "memory-engine/lib/me-traits",
     "memory-engine/lib/me-storage",
+    "memory-engine/lib/me-index",
     "memory-engine/lib/memory-engine",
     "memory-engine/lib/embed",
     "memory-engine/bin/cli",
