+++ memory-engine/lib/me-index/src/graph/mod.rs
@@ -3,7 +3,7 @@
 //! Node weights are fact ids; edge weights are [`EdgeData`]. Only active
 //! (non-expired) edges are represented, and the graph is edge-only — isolated
 //! facts with no active edge are not materialized as nodes (matching
-//! [`MemoryGraph::load_from_db`] semantics).
+//! [`MemoryGraph::from_active_edges`] semantics).
 
 mod memory_graph;
